requirements.txt
Flask
boto3
gunicorn
python-dotenv
beautifulsoup4
spacy